Solve App issues
